Reorg component directory structure
