+++ src/App.js
@@ -9,8 +9,8 @@ import ReactGA from "react-ga";
 //Import Components
 import Navigation from "./components/layout/Navigation";
 import Hero from "./components/layout/Hero";
-import About from "./components/layout/AboutSection/About";
-import Instances from "./components/layout/Instances";
+import About from "./components/layout/About/About";
+import Instances from "./components/layout/Instances/Instances";
 import Contributing from "./components/layout/Contributing";
 import Footer from "./components/layout/Footer";
 
@@ -21,7 +21,7 @@ function App() {
   return (
     <div id="top" className="App">
       <Navigation
-        NavLinks={[
+        links={[
           {
             name: "About",
             to: "#about",
